+++ packages/icons/src/exports/Bolt.tsx
@@ -1,0 +1,32 @@
+/* THIS FILE IS AUTO-GENERATED – DO NOT EDIT */
+
+import {forwardRef, type ForwardRefExoticComponent, type RefAttributes, type SVGProps} from 'react'
+
+/**
+ * @public
+ */
+export const BoltIcon: ForwardRefExoticComponent<
+  Omit<SVGProps<SVGSVGElement>, 'ref'> & RefAttributes<SVGSVGElement>
+> = /* @__PURE__ */ forwardRef(function BoltIcon(props, ref) {
+  return (
+    <svg
+      data-sanity-icon="bolt"
+      width="1em"
+      height="1em"
+      viewBox="0 0 25 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+      ref={ref}
+    >
+      <path
+        d="M18 11.5L9 21L11 13.5H7L16 4L14 11.5H18Z"
+        stroke="currentColor"
+        strokeWidth={1.2}
+        strokeLinejoin="round"
+      />
+    </svg>
+  )
+})
+
+export {BoltIcon as default}