+++ packages/icons/src/exports/Checkmark.tsx
@@ -1,0 +1,32 @@
+/* THIS FILE IS AUTO-GENERATED – DO NOT EDIT */
+
+import {forwardRef, type ForwardRefExoticComponent, type RefAttributes, type SVGProps} from 'react'
+
+/**
+ * @public
+ */
+export const CheckmarkIcon: ForwardRefExoticComponent<
+  Omit<SVGProps<SVGSVGElement>, 'ref'> & RefAttributes<SVGSVGElement>
+> = /* @__PURE__ */ forwardRef(function CheckmarkIcon(props, ref) {
+  return (
+    <svg
+      data-sanity-icon="checkmark"
+      width="1em"
+      height="1em"
+      viewBox="0 0 25 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+      ref={ref}
+    >
+      <path
+        d="M5.5 11.5L10.5 16.5L19.5 7.60001"
+        stroke="currentColor"
+        strokeWidth={1.2}
+        strokeLinejoin="round"
+      />
+    </svg>
+  )
+})
+
+export {CheckmarkIcon as default}